requirements.txt
discord.py>=2.4.0
python-dotenv>=1.0.1
aiohttp>=3.9.5
edge-tts>=6.1.12
PyNaCl>=1.5.0
discord-ext-voice-recv>=0.5.2a179
